backend/djangoapp/app/requirements.txt
aniso8601==9.0.1
argcomplete==3.1.1
asgiref==3.8.1
click==8.1.7
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
Markdown==3.6
packaging==23.1
pipx==1.2.0
promise==2.3
psycopg2==2.9.9
PyYAML @ file:///private/tmp/pyyaml-20230717-6270-1iiwi1j/PyYAML-6.0.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
userpath==1.9.0
